use integer tabwidth and follow normal c conventions

+++ src/cmd/samterm/samterm.h
@@ -40,7 +40,7 @@ struct Text
 	ushort	tag;
 	char	lock;
 	Flayer	l[NL];		/* screen storage */
-	long	tabwidth;
+	int	tabwidth;
 };
 
 struct Readbuf
